Split AccountProcessor
